✨ feat(Commons): refactor components to use new directory structure for better organization and maintainability
