+++ src/app/(main-site)/contact/page.tsx
@@ -1,5 +1,5 @@
-import CtaDevis from '@/components/CtaDevis';
 import HeroHeader from '@/components/commons/HeroHeader';
+import CtaDevis from '@/components/main-site/CtaDevis';
 import { Metadata } from 'next';
 import { useTranslations } from 'next-intl';
 import Image from 'next/image';
