+++ src/app/(main-site)/page.tsx
@@ -1,5 +1,5 @@
-import CtaContact from '@/components/CtaContact';
-import ServiceCard from '@/components/ServiceCard';
+import CtaContact from '@/components/main-site/CtaContact';
+import ServiceCard from '@/components/main-site/ServiceCard';
 import { Metadata } from 'next';
 import { getTranslations } from 'next-intl/server';
 import Image from 'next/image';
