+++ src/components/commons/FooterIndicator.tsx
